refactor: site config
