+++ src/app/robots.ts
@@ -1,6 +1,6 @@
 import { MetadataRoute } from 'next';
 
-import { siteConfig } from '@/config/site';
+import { siteConfig } from '@/lib/config';
 
 export default function robots(): MetadataRoute.Robots {
   const baseUrl = process.env.NEXT_PUBLIC_APP_URL || siteConfig.url;
