+++ src/components/layout/site-footer.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 
-import { siteConfig } from '@/config/site';
+import { siteConfig } from '@/lib/config';
 import { cn } from '@/lib/utils';
 
 import { buttonVariants } from '@/components/shadcn-ui/button';
