+++ src/lib/config.ts
@@ -1,14 +1,7 @@
-import { SiteConfig } from '@/types';
-
-/**
- * TODO: get site_url via T3 env
- *
- * @see https://env.t3.gg/
- */
-export const siteConfig: SiteConfig = {
+export const siteConfig = {
   name: 'next-minimal-blog',
   description:
-    'next-minimal-blog is a lightweight, minimalistic blog template built with Next.js 15, MDX, Tailwind CSS, and shadcn/ui',
+    'next-minimal-blog is a lightweight, minimalistic blog template built with Next.js (v15), MDX, Tailwind CSS (v4), and shadcn/ui',
   url: 'https://next-minimal-blog-delta.vercel.app/',
   ogImage: 'https://next-minimal-blog-delta.vercel.app/og.png',
   links: {
